+++ resources/views/form-input/dedication-output-patent.blade.php
@@ -34,7 +34,8 @@ $errors->has('patent_type') || old('patent_type')
                     <div class="form-group">
                         <label for="patent_no" class="col-sm-4 col-md-3 control-label">Nomor SK Paten</label>
                         <div class="col-sm-7">
-                            <input name="patent_no" class="form-control input-sm" type="text" value="{{ $dedication_output_patent->patent_no }}">
+                            <input name="patent_no" class="form-control input-sm" type="text"
+                                   value="{{ $dedication_output_patent->patent_no }}" {{$upd_mode !== 'approve' ? '' : 'disabled'}}>
                             @if($errors->has('patent_no'))
                                 <label class="error" for="patent_no" style="display: inline-block;">
                                     {{ $errors->first('patent_no') }}
@@ -46,7 +47,8 @@ $errors->has('patent_type') || old('patent_type')
                     <div class="form-group">
                         <label for="patent_year" class="col-sm-4 col-md-3 control-label">Tahun Paten</label>
                         <div class="col-sm-7">
-                            <input name="patent_year" class="form-control input-sm" type="text" value="{{ $dedication_output_patent->patent_year }}">
+                            <input name="patent_year" class="form-control input-sm" type="text"
+                                   value="{{ $dedication_output_patent->patent_year }}" {{$upd_mode !== 'approve' ? '' : 'disabled'}}>
                             @if($errors->has('patent_year'))
                                 <label class="error" for="patent_year" style="display: inline-block;">
                                     {{ $errors->first('patent_year') }}
@@ -56,9 +58,11 @@ $errors->has('patent_type') || old('patent_type')
                     </div>
 
                     <div class="form-group">
-                        <label for="patent_owner_name" class="col-sm-4 col-md-3 control-label">Nama Pemilik Paten</label>
+                        <label for="patent_owner_name" class="col-sm-4 col-md-3 control-label">Nama Pemilik
+                            Paten</label>
                         <div class="col-sm-7">
-                            <input name="patent_owner_name" class="form-control input-sm" type="text" value="{{ $dedication_output_patent->patent_owner_name }}">
+                            <input name="patent_owner_name" class="form-control input-sm" type="text"
+                                   value="{{ $dedication_output_patent->patent_owner_name }}" {{$upd_mode !== 'approve' ? '' : 'disabled'}}>
                             @if($errors->has('patent_owner_name'))
                                 <label class="error" for="patent_owner_name" style="display: inline-block;">
                                     {{ $errors->first('patent_owner_name') }}
@@ -70,7 +74,8 @@ $errors->has('patent_type') || old('patent_type')
                     <div class="form-group">
                         <label for="patent_type" class="col-sm-4 col-md-3 control-label">Jenis Paten</label>
                         <div class="col-sm-7">
-                            <input name="patent_type" class="form-control input-sm" type="text" value="{{ $dedication_output_patent->patent_type }}">
+                            <input name="patent_type" class="form-control input-sm" type="text"
+                                   value="{{ $dedication_output_patent->patent_type }}" {{$upd_mode !== 'approve' ? '' : 'disabled'}}>
                             @if($errors->has('patent_type'))
                                 <label class="error" for="patent_type" style="display: inline-block;">
                                     {{ $errors->first('patent_type') }}
@@ -96,42 +101,46 @@ $errors->has('patent_type') || old('patent_type')
                             </div>
                         @endif
 
-                        <div class="clearfix"></div>
-                        <label class="control-label col-sm-4 col-md-3">Unggah Sertifikat Paten</label>
-                        <div class="col-sm-7">
-                            <div class="fileinput fileinput-new input-group" data-provides="fileinput">
-                                <div class="form-control input-sm" data-trigger="fileinput">
-                                    <i class="glyphicon glyphicon-file fileinput-exists"></i>
-                                    <span class="fileinput-filename"></span>
-                                </div>
+                        @if($upd_mode !== 'approve')
+                            <div class="clearfix"></div>
+                            <label class="control-label col-sm-4 col-md-3">Unggah Sertifikat Paten</label>
+                            <div class="col-sm-7">
+                                <div class="fileinput fileinput-new input-group" data-provides="fileinput">
+                                    <div class="form-control input-sm" data-trigger="fileinput">
+                                        <i class="glyphicon glyphicon-file fileinput-exists"></i>
+                                        <span class="fileinput-filename"></span>
+                                    </div>
                                             <span class="input-group-addon btn btn-success btn-file">
                                                 <span class="fileinput-new">Select file</span>
                                                 <span class="fileinput-exists">Change</span>
                                                 <input type="file" name="file_patent"
                                                        value="">
                                             </span>
-                                <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
-                                   data-dismiss="fileinput">Remove</a>
+                                    <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
+                                       data-dismiss="fileinput">Remove</a>
+                                </div>
+                                @if($errors->has('file_patent'))
+                                    <label class="error" for="file_patent"
+                                           style="display: inline-block;">
+                                        {{ $errors->first('file_patent') }}
+                                    </label>
+                                @endif
                             </div>
-                            @if($errors->has('file_patent'))
-                                <label class="error" for="file_patent"
-                                       style="display: inline-block;">
-                                    {{ $errors->first('file_patent') }}
-                                </label>
-                            @endif
-                        </div>
+                        @endif
                     </div> <!-- /.form-group -->
 
-                    {{ csrf_field() }}
-                    <input type="hidden" name="_method" value="PUT">
+                    @if($upd_mode !== 'approve')
+                        {{ csrf_field() }}
+                        <input type="hidden" name="_method" value="PUT">
 
-                    <div class="clearfix"></div>
-                    <div class="form-footer">
-                        <div class="col-sm-offset-4 col-md-offset-3">
-                            <a href="{{url($deleteUrl)}}" class="btn btn-danger btn-slideright">Kembali</a>
-                            <button type="submit" class="btn btn-success btn-slideright">Submit</button>
-                        </div><!-- /.col-sm-offset-3 -->
-                    </div><!-- /.form-footer -->
+                        <div class="clearfix"></div>
+                        <div class="form-footer">
+                            <div class="col-sm-offset-4 col-md-offset-3">
+                                <a href="{{url($deleteUrl)}}" class="btn btn-danger btn-slideright">Kembali</a>
+                                <button type="submit" class="btn btn-success btn-slideright">Submit</button>
+                            </div><!-- /.col-sm-offset-3 -->
+                        </div><!-- /.form-footer -->
+                    @endif
                 </form>
             </div>
         </div>
