+++ resources/views/form-input/dedication-output-product.blade.php
@@ -42,30 +42,32 @@
                             </div>
                         @endif
 
-                        <div class="clearfix"></div>
-                        <label class="control-label col-sm-4 col-md-3">Unggah Blueprint</label>
-                        <div class="col-sm-7">
-                            <div class="fileinput fileinput-new input-group" data-provides="fileinput">
-                                <div class="form-control input-sm" data-trigger="fileinput">
-                                    <i class="glyphicon glyphicon-file fileinput-exists"></i>
-                                    <span class="fileinput-filename"></span>
+                        @if($upd_mode !== 'approve')
+                            <div class="clearfix"></div>
+                            <label class="control-label col-sm-4 col-md-3">Unggah Blueprint</label>
+                            <div class="col-sm-7">
+                                <div class="fileinput fileinput-new input-group" data-provides="fileinput">
+                                    <div class="form-control input-sm" data-trigger="fileinput">
+                                        <i class="glyphicon glyphicon-file fileinput-exists"></i>
+                                        <span class="fileinput-filename"></span>
+                                    </div>
+                                    <span class="input-group-addon btn btn-success btn-file">
+                                        <span class="fileinput-new">Select file</span>
+                                        <span class="fileinput-exists">Change</span>
+                                        <input type="file" name="file_blueprint"
+                                               value="">
+                                    </span>
+                                    <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
+                                       data-dismiss="fileinput">Remove</a>
                                 </div>
-                                <span class="input-group-addon btn btn-success btn-file">
-                                    <span class="fileinput-new">Select file</span>
-                                    <span class="fileinput-exists">Change</span>
-                                    <input type="file" name="file_blueprint"
-                                           value="">
-                                </span>
-                                <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
-                                   data-dismiss="fileinput">Remove</a>
+                                @if($errors->has('file_blueprint'))
+                                    <label class="error" for="file_blueprint"
+                                           style="display: inline-block;">
+                                        {{ $errors->first('file_blueprint') }}
+                                    </label>
+                                @endif
                             </div>
-                            @if($errors->has('file_blueprint'))
-                                <label class="error" for="file_blueprint"
-                                       style="display: inline-block;">
-                                    {{ $errors->first('file_blueprint') }}
-                                </label>
-                            @endif
-                        </div>
+                        @endif
                     </div> <!-- /.form-group -->
 
                     <div class="form-group">
@@ -85,30 +87,32 @@
                             </div>
                         @endif
 
-                        <div class="clearfix"></div>
-                        <label class="control-label col-sm-4 col-md-3">Unggah Barang Jadi</label>
-                        <div class="col-sm-7">
-                            <div class="fileinput fileinput-new input-group" data-provides="fileinput">
-                                <div class="form-control input-sm" data-trigger="fileinput">
-                                    <i class="glyphicon glyphicon-file fileinput-exists"></i>
-                                    <span class="fileinput-filename"></span>
-                                </div>
+                        @if($upd_mode !== 'approve')
+                            <div class="clearfix"></div>
+                            <label class="control-label col-sm-4 col-md-3">Unggah Barang Jadi</label>
+                            <div class="col-sm-7">
+                                <div class="fileinput fileinput-new input-group" data-provides="fileinput">
+                                    <div class="form-control input-sm" data-trigger="fileinput">
+                                        <i class="glyphicon glyphicon-file fileinput-exists"></i>
+                                        <span class="fileinput-filename"></span>
+                                    </div>
                                 <span class="input-group-addon btn btn-success btn-file">
                                     <span class="fileinput-new">Select file</span>
                                     <span class="fileinput-exists">Change</span>
                                     <input type="file" name="file_finished_good"
                                            value="">
                                 </span>
-                                <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
-                                   data-dismiss="fileinput">Remove</a>
+                                    <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
+                                       data-dismiss="fileinput">Remove</a>
+                                </div>
+                                @if($errors->has('file_finished_good'))
+                                    <label class="error" for="file_finished_good"
+                                           style="display: inline-block;">
+                                        {{ $errors->first('file_finished_good') }}
+                                    </label>
+                                @endif
                             </div>
-                            @if($errors->has('file_finished_good'))
-                                <label class="error" for="file_finished_good"
-                                       style="display: inline-block;">
-                                    {{ $errors->first('file_finished_good') }}
-                                </label>
-                            @endif
-                        </div>
+                        @endif
                     </div> <!-- /.form-group -->
 
                     <div class="form-group">
@@ -128,42 +132,46 @@
                             </div>
                         @endif
 
-                        <div class="clearfix"></div>
-                        <label class="control-label col-sm-4 col-md-3">Unggah Gambar Kerja</label>
-                        <div class="col-sm-7">
-                            <div class="fileinput fileinput-new input-group" data-provides="fileinput">
-                                <div class="form-control input-sm" data-trigger="fileinput">
-                                    <i class="glyphicon glyphicon-file fileinput-exists"></i>
-                                    <span class="fileinput-filename"></span>
-                                </div>
+                        @if($upd_mode !== 'approve')
+                            <div class="clearfix"></div>
+                            <label class="control-label col-sm-4 col-md-3">Unggah Gambar Kerja</label>
+                            <div class="col-sm-7">
+                                <div class="fileinput fileinput-new input-group" data-provides="fileinput">
+                                    <div class="form-control input-sm" data-trigger="fileinput">
+                                        <i class="glyphicon glyphicon-file fileinput-exists"></i>
+                                        <span class="fileinput-filename"></span>
+                                    </div>
                                 <span class="input-group-addon btn btn-success btn-file">
                                     <span class="fileinput-new">Select file</span>
                                     <span class="fileinput-exists">Change</span>
                                     <input type="file" name="file_working_pic"
                                            value="">
                                 </span>
-                                <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
-                                   data-dismiss="fileinput">Remove</a>
+                                    <a href="#" class="input-group-addon btn btn-danger fileinput-exists"
+                                       data-dismiss="fileinput">Remove</a>
+                                </div>
+                                @if($errors->has('file_working_pic'))
+                                    <label class="error" for="file_working_pic"
+                                           style="display: inline-block;">
+                                        {{ $errors->first('file_working_pic') }}
+                                    </label>
+                                @endif
                             </div>
-                            @if($errors->has('file_working_pic'))
-                                <label class="error" for="file_working_pic"
-                                       style="display: inline-block;">
-                                    {{ $errors->first('file_working_pic') }}
-                                </label>
-                            @endif
-                        </div>
+                        @endif
                     </div> <!-- /.form-group -->
 
-                    {{ csrf_field() }}
-                    <input type="hidden" name="_method" value="PUT">
+                    @if($upd_mode !== 'approve')
+                        {{ csrf_field() }}
+                        <input type="hidden" name="_method" value="PUT">
 
-                    <div class="clearfix"></div>
-                    <div class="form-footer">
-                        <div class="col-sm-offset-4 col-md-offset-3">
-                            <a href="{{url($deleteUrl)}}" class="btn btn-danger btn-slideright">Kembali</a>
-                            <button type="submit" class="btn btn-success btn-slideright">Submit</button>
-                        </div><!-- /.col-sm-offset-3 -->
-                    </div><!-- /.form-footer -->
+                        <div class="clearfix"></div>
+                        <div class="form-footer">
+                            <div class="col-sm-offset-4 col-md-offset-3">
+                                <a href="{{url($deleteUrl)}}" class="btn btn-danger btn-slideright">Kembali</a>
+                                <button type="submit" class="btn btn-success btn-slideright">Submit</button>
+                            </div><!-- /.col-sm-offset-3 -->
+                        </div><!-- /.form-footer -->
+                    @endif
                 </form>
             </div>
         </div>
